fix(script): fix proto script, modify venv script

+++ scripts/start_venv.sh
@@ -1,8 +1,17 @@
 #!/bin/sh
 
-rm -rf lib
-rm -rf bin
-rm -rf pyvenv.cfg
+
+if [ -d "lib" ]; then
+    rm -rf lib
+fi
+
+if [ -d "bin" ]; then
+    rm -rf bin
+fi
+
+if [ -f "pyvenv.cfg" ]; then
+    rm -rf pyvenv.cfg
+fi
 
 python3 -m venv ./
 source ./bin/activate
